Se re-acomodan archivos html en las aplicaciones + title + se ajusta diseño con css en varios html
