+++ static/css/appves/biomas.css
@@ -1,0 +1,19 @@
+h1 {
+  text-align: center;
+  margin-bottom: 20px;
+}
+
+img.center {
+  display: block;
+  margin-left: auto;
+  margin-right: auto;
+  width: 1100px;
+  border-radius: 25px;
+}
+
+.ul-block {
+  list-style: none;
+  padding: 0;
+  column-count: 3;
+  column-gap: 150px;
+}
