+++ static/css/appves/listar_aves.css
@@ -1,3 +1,8 @@
+h1 {
+  text-align: center;
+  margin-bottom: 20px;
+}
+
 .row-cols-1 {
   display: flex;
   flex-wrap: wrap;
