+++ static/css/appves/topografia.css
@@ -1,0 +1,11 @@
+h1 {
+    text-align: center;
+    margin-bottom: 20px;
+}
+
+img.center {
+    display: block;
+    margin-left: auto;
+    margin-right: auto;
+    width: 1100px;
+}
