+++ templates/appves/about.html
@@ -1,4 +1,9 @@
 {% extends 'base.html' %}
+{% block title %}
+<title>Sobre Appves</title>
+{% endblock %}
+
+{% load static %}
 
 {% block content %}
