+++ templates/appves/glosario.html
@@ -1,5 +1,7 @@
 {% extends 'base.html' %}
-
+{% block title %}
+<title>Glosario</title>
+{% endblock %}
 {% block content %}
 
 <h1>Glosario</h1>
