+++ templates/appves/home.html
@@ -1,0 +1,8 @@
+{% extends 'base.html' %}
+{% block title %}
+<title>Home</title>
+{% endblock %}
+
+{% block content %}
+<h1>Appves</h1>
+{% endblock %}
