+++ templates/appves/lugares.html
@@ -1,0 +1,12 @@
+{% extends 'base.html' %}
+{% block title %}
+<title>Lugares para observar</title>
+{% endblock %}
+
+{% block content %}
+
+<h1>Lugares para observar aves</h1>
+
+{% lorem 9 b random %}
+
+{% endblock %}
