+++ templates/appves/topografia.html
@@ -1,0 +1,15 @@
+{% extends 'base.html' %}
+{% block title %}
+<title>Topografía de Aves</title>
+{% endblock %}
+{% load static %}
+
+{% block content %}
+{% block styles %}
+  <link rel="stylesheet" href="{% static 'css/appves/topografia.css' %}">
+{% endblock %}
+<h1>Topografía de las aves</h1>
+
+<img class="center" src="https://i.pinimg.com/originals/f1/71/7a/f1717af1f4e544c3ab09623fbbcae6f8.jpg" alt="topografia de las aves">
+
+{% endblock %}
