--- templates/lugares.html
@@ -1,7 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-
-{% lorem 9 b random %}
-
-{% endblock %}
