--- templates/quienes.html
@@ -1,17 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-
-<h1>Quiénes somos</h1>
-
-<p>Somos un grupo de compañeros de la Tecnicatura en Análisis de Sistemas. Carrera que se dicta en Instituto de Formación Técnica Superior N° 18 </p>
-
-
-<p>Esta es nuestro trabajo final.</p>
-
-<ul>
-    <li>Manuel Jigena</li>
-    <li>Gonzalo Roget</li>
-    <li>Fernando Warno</li>
-</ul>
-{% endblock %}
