--- templates/topografia.html
@@ -1,9 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-
-<h1>Topografía de las aves</h1>
-
-<img src="https://i.pinimg.com/originals/f1/71/7a/f1717af1f4e544c3ab09623fbbcae6f8.jpg" alt="topografia de las aves">
-
-{% endblock %}
