+++ templates/user/create_user.html
@@ -1,5 +1,7 @@
 {% extends 'base.html' %}
-
+{% block title %}
+<title>Crear Cuenta</title>
+{% endblock %}
 {% block content %}
 <h1>Crear cuenta</h1>
