+++ templates/user/login_user.html
@@ -3,7 +3,9 @@
 {% load static %}
 
 {% block content %}
-
+{% block title %}
+<title>Login</title>
+{% endblock %}
 {% block styles %}
 <link rel="stylesheet" href="{% static 'css/user/login_user.css' %}">
 {% endblock %}
