+++ templates/user/quienes.html
@@ -1,0 +1,20 @@
+{% extends 'base.html' %}
+{% block title %}
+<title>Quiénes somos?</title>
+{% endblock %}
+{% block content %}
+
+<h1>Quiénes somos</h1>
+
+<p>Somos un grupo de compañeros de la Tecnicatura en Análisis de Sistemas.</p>
+<p>Carrera que se dicta en Instituto de Formación Técnica Superior N° 18.</p>
+
+
+<p>Este aplicativo Web es nuestro trabajo final.</p>
+
+<ul>
+    <li>Manuel Jigena</li>
+    <li>Gonzalo Roget</li>
+    <li>Fernando Warno</li>
+</ul>
+{% endblock %}
