+++ user/views.py
@@ -7,29 +7,21 @@ from django.contrib.auth.decorators import login_required
 
 # Create your views here.
 # @login_required
-def home(request):
-    return render(request, 'home.html')
 
-def about(request):
-    return render(request, 'about.html')
 
-def biomas(request):
-    return render(request, 'biomas.html')
 
-def glosario(request):
-    return render(request, 'glosario.html')
 
-def lugares(request):
-    return render(request, 'lugares.html')
 
-def topografia(request):
-    return render(request, 'topografia.html')
 
-def familias(request):
-    return render(request, 'familias.html')
+
+
+
 
 def quienes(request):
-    return render(request, 'quienes.html')
+    return render(request, 'user/quienes.html')
+
+
+
 
 def crear_cuenta(request):
     if request.method == 'GET':
